Fix a bug in KeyStore, and add a regression test.
When given a key to open an existing KeyStore, we iterate through each
wrapped key and try to unwrap it with the given key. Previously, we
try!()'ed this operation, so if we encountered some other key first we
would stop immediately, instead of ignoring the error and trying the
remaining keys.

The net result of this was that if you created a KeyStore with one key,
and then added a new one, the new key would incorrectly fail to open the
existing KeyStore.