Get rid of SensitiveData.
The idea of trying to zero out memory was kind of fine, in that it was
explained that it wasn't guaranteed, but doing so was making other
things more complicated and it got to the point where there were too
many copies or references we had to spread around anyway, so it's no
longer worth it.

Bottom line is, if you decrypt data on your machine, and an adversary
can read arbitrary data from RAM, then you're pwned anyway.