Add LazyResult, which wraps Lazy and provides better error handling.
The immediate benefit of this is, when we try to open our keystore in
repository.rs, and we get an error, RUST_BACKTRACE now shows the actual
origin of the error because we don't create a new error instance in
repository.rs.

This was done previously to work around the fact that Lazy returned us
an &Error, instead of an Error.