Refactor to get ready to add an "rmpiv" command.
Add a placeholder for the new command, with no implementation yet.

Refactor prompt_for_reader so it's more readable, and more generally
useful - now we can use the same logic to pick from any list of devices
(e.g., the list of devices in our crypto configuration), not just the
list returned by `list_readers`.

Also, it now calls out serial numbers explicitly, as those are the
identifiers we're primarily relying on.