Add a first "rmpiv" command implementation.
This implementation works, but there is a significant oversight: as
written, the user has to unlock the keystore before removing the key.
This is awkward, as it means they may be prompted to unlock using the
key they're trying to remove.

KeyStore itself doesn't require this, it's just a product of how `pwm`
is written currently. This will be fixed in a future commit.