Fix a key store bug when opening with a hard-coded password.
In the case of a hard-coded password, we don't loop asking for a new
password prompt. But, in this case, we need to return an error, rather
than just printing a warning and returning "Ok".

The result of this is that we were were sometimes returning an "Ok"
key store, which was actually not "opened" yet and thus was unusable.