Fix tests failing when no smartcards are present on the system.
It turns out simply establishing a PCSC context requires communicating
with smart card devices? Or maybe just the PCSC daemon running on the
system? And so, on systems with no smart cards, no daemon, etc., this
fails.

Previously, we'd bail out entirely on this error. But, it's perfectly
reasonable for a system to not have any of this infrastructure. In this
case, just return an empty list of PCSC readers, so we fall back to
password authentication.