Default the piv feature off, since it is still experimental.
This "next" branch contains lots of useful fixes and improvements, but
it also contains the piv / yubikey feature stuff. That feature has taken
much longer than expected to finish / stabilize, but in the meantime, I
want the rest of the fixes / improvements in master!

So, just mark that feature off by default, to indicate that it's still
experimental, and then we can just about merge the rest of the stuff
into master.