Don't implement Serialize/Deserialize for Secret.
In general we don't want to be serializing secrets out. It turns out the
only user of this was when we were serializing out *ciphertext*. Since
the data is already encrypted, it's no longer secret. So we can just not
use Secret for this at all, and remove the implementation.