Pad with zeros instead of random bytes for development builds.
We pad *before encrypting*, so I think the actual bytes we pad with
don't matter. (Assuming our cipher is good, nobody should be able to
deduce things about the plaintext anyway.)

The tradeoff here is, this makes our encryption deterministic. For
testing this is really useful, because output is reproducible.