Switch from thiserror to anyhow.
Why? Because thiserror doesn't seem to have full backtrace support
except on nightly Rust. This makes debugging difficult.

Also, thiserror has some overhead: it requires a fair amount of
boilerplate, both in defining the error types and in returning new
errors. This is really an application, not a library re-used by many
applications, so we're basically paying all of those costs without
getting many of the benefits.