Switch to newer bdrck, and away from sodiumoxide.
sodiumoxide is no longer maintained, and it has known security issues.
So we want to get rid of that dependency.

To do so, we just switch to a newer bdrck version. This version is based
on our own halite-sys crate, which is a direct binding of upstream NaCl.

Newer bdrck also provides its own Secret type, so we can delete ours and
switch to that. This is unfortunately a little bit involved, despite our
prep work, because there are some lingering differences between the old
and new types.

Also, bdrck started enforcing that init() is called before crypto
functions can be used (debug builds only). So, we need to update all of
our test cases to also call our own init function. This is something we
should have been doing anyway, and which was overlooked before.